+++ lib/src/element/super/spwml_element.dart
@@ -316,6 +316,28 @@ class SpWMLElement extends StatelessWidget {
     throw UnimplementedError();
   }
 
+  /// (en) Gets the ID set for this element. Returns null if not set.
+  ///
+  /// (ja) このエレメントに設定されているIDを取得します。設定されていない場合はnullを返します。
+  int? getID() {
+    if (params.containsKey(EnumSpWMLParams.id)) {
+      return int.tryParse(params[EnumSpWMLParams.id]);
+    } else {
+      return null;
+    }
+  }
+
+  /// (en) Gets the SID set for this element. Returns null if not set.
+  ///
+  /// (ja) このエレメントに設定されているSIDを取得します。設定されていない場合はnullを返します。
+  String? getSID() {
+    if (params.containsKey(EnumSpWMLParams.sid)) {
+      return params[EnumSpWMLParams.sid];
+    } else {
+      return null;
+    }
+  }
+
   /// Wrap if necessary.
   /// * [child] : The child widget.
   @protected